+++ src/main/java/it/sevenbits/calculator/operations/Operation.java
@@ -3,7 +3,7 @@ package it.sevenbits.calculator.operations;
 public interface Operation {
     String getSymbol();
 
-    float apply(float... operands);
+    double apply(double... operands);
 
     int getArity();
 }
